requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
filetype==1.2.0
fonttools==4.55.3
idna==3.7
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.50
six==1.17.0
tqdm==4.67.1
urllib3==2.3.0
